WIP: move functionality to feature module
