+++ kotlin/features/authentication/feature-authentication-core/build.gradle
@@ -11,6 +11,6 @@ dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
     api project(path: ':kotlin:concepts:authentication:concept-authentication')
     api project(path: ':kotlin:concepts:concept-coroutines')
-    api project(path: ':kotlin:concepts:concept-foreground-state')
+    api project(path: ':kotlin:features:feature-foreground-state')
     implementation project(path: ':kotlin:crypto:k-openssl')
 }
