+++ kotlin/features/feature-foreground-state/.gitignore
@@ -1,0 +1,1 @@
+/build
