+++ kotlin/features/feature-foreground-state/build.gradle
@@ -1,0 +1,13 @@
+plugins {
+    id 'java-library'
+    id 'kotlin'
+}
+
+sourceCompatibility = JavaVersion.VERSION_1_8
+targetCompatibility = JavaVersion.VERSION_1_8
+
+dependencies {
+    implementation fileTree(dir: "libs", include: ["*.jar"])
+    api project(path: ':kotlin:concepts:concept-foreground-state')
+    implementation deps.kotlin.coroutinesCore
+}
